requirements.txt
asgiref==3.5.2
astroid==2.4.2
dj-database-url==1.0.0
Django==3.2.10
django-filter==22.1
django-on-heroku==1.1.2
djangorestframework==3.11.0
Faker==15.0.0
gunicorn==20.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.9.4
pylint==2.5.3
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.5.4
validate-docbr==1.10.0
whitenoise==6.2.0
wrapt==1.12.1
